+++ presentation/src/main/java/com/semicolon/walkhub/ui/profile/setting/ui/model/ThirdSearchSchoolData.kt
@@ -4,7 +4,7 @@ data class ThirdSearchSchoolData(
     val schoolList: List<SchoolInfo>
 ) {
     data class SchoolInfo(
-        val schoolId: Int,
+        val schoolId: Long,
         val logoImageUrl: String,
         val schoolName: String
     )
